backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
ultralytics>=8.4
pillow>=10.0
